Smarter input file handling
This PR refactors the `get_files` function to behave more modularly.  Reruns are
triggered only for checkpoints that are affected by changes to the input
arguments.  The underlying code separates the file search into several
internal functions (`_get_sim_files`, `_get_data_files`, etc) to make it
easier to see which input arguments are associated with which checkpoints
further down the pipeline.

This code has been tested on the example dataset in the repository.