Smarter input file handling (#23)
This PR refactors the `get_files` function to behave more modularly.  Reruns are triggered only for checkpoints that are affected by changes to the input arguments.  The underlying code separates the file search into several internal functions (`_get_sim_files`, `_get_data_files`, etc) to make it easier to see which input arguments are associated with which checkpoints further down the pipeline.  Backwards-compatible with existing `files.npz` checkpoint files.

The options `subtract_template_noise` and `subtract_reference_signal` are removed, and instead controlled by setting the `template_noise_type` and `reference_type` variables, respectively.  Note that this changes the file structure for reference maps used for debiasing null tests.

This code has been tested on the example dataset in the repository, and on the SPIDER publication dataset.

Fixes #20